+++ packages/core/src/components/Avatar/Avatar.test.tsx
@@ -5,7 +5,7 @@ import { Avatar } from './Avatar';
 
 describe('Avatar component', () => {
     it('should render with default theme', () => {
-        const { container } = render(<Avatar>M</Avatar>);
+        const { container } = render(<Avatar withHoverEffect={true}>M</Avatar>);
         expect(container).toMatchSnapshot();
     });
 
@@ -15,6 +15,7 @@ describe('Avatar component', () => {
                 size="M"
                 textColor={defaultTheme.colors.green[500]}
                 bgColor={defaultTheme.colors.green[100]}
+                withHoverEffect={true}
                 hoverBgColor={defaultTheme.colors.green[400]}
                 hoverTextColor={defaultTheme.colors.white}
                 hoverImgShadowColor="rgba(96, 120, 144, 0.35)"
@@ -38,8 +39,26 @@ describe('Avatar component', () => {
 
     it('should render image avatar properly', () => {
         const { container } = render(
-            <Avatar size="L">
+            <Avatar size="L" withHoverEffect={true}>
                 <img src="http://dummurl" />
+            </Avatar>
+        );
+        expect(container).toMatchSnapshot();
+    });
+
+    it('should render avatar without hover', () => {
+        const { container } = render(
+            <Avatar
+                size="M"
+                textColor={defaultTheme.colors.green[500]}
+                bgColor={defaultTheme.colors.green[100]}
+                hoverBgColor={defaultTheme.colors.green[400]}
+                hoverTextColor={defaultTheme.colors.white}
+                hoverImgShadowColor="rgba(96, 120, 144, 0.35)"
+                hoverTextShadowColor="rgba(0, 128, 0, 0.35)"
+                onClick={jest.fn()}
+            >
+                M
             </Avatar>
         );
         expect(container).toMatchSnapshot();