+++ packages/core/src/components/Radio/Radio.styled.tsx
@@ -1,49 +1,69 @@
-import { defaultTheme } from '@medly-components/theme';
-import { centerAligned, styled, WithThemeProp } from '@medly-components/utils';
-import { Props, RadioWrapperProps } from './types';
+import { RadioSizes } from '@medly-components/theme';
+import { styled, WithThemeProp } from '@medly-components/utils';
+import { getSelectorLabelPositionStyle } from '../Selectors';
+import { Props, WrapperProps } from './types';
 
-export const RadioFillStyled = styled('div')<WithThemeProp>`
-    background-color: ${({ theme }) => theme.radio.fillColor};
-    width: 0;
-    height: 0;
-    z-index: 1;
-    border-radius: 100%;
-    ${centerAligned()}
-`;
-RadioFillStyled.defaultProps = {
-    theme: defaultTheme
-};
+const getRadioSize = ({ theme, size }: { size?: RadioSizes } & WithThemeProp) => theme.radio.sizes[size || theme.radio.defaultSize];
 
-export const RadioStyled = styled('input').attrs({ type: 'radio' })<Props>`
-    opacity: 0;
-    z-index: 2;
-    position: absolute;
+export const StyledRadio = styled('div')`
+    transition: all 100ms ease-out;
+    box-sizing: border-box;
+    border: 0.15rem solid;
     width: 100%;
     height: 100%;
-    margin: 0;
-    cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
+    border-radius: 100%;
+`;
 
-    &:checked ~ ${RadioFillStyled} {
-        width: calc(75%);
-        height: calc(75%);
-        transition: width 0.2s ease, height 0.2s ease;
+export const HiddenRadio = styled('input').attrs(({ theme }) => ({ type: 'radio', ...theme.radio }))<Props>`
+    position: absolute;
+    opacity: 0;
+    width: 1%;
+    height: 1%;
+    margin: 0;
+
+    & ~ ${StyledRadio} {
+        border-color: ${({ borderColor, hasError }) => borderColor[hasError ? 'error' : 'default']};
+    }
+
+    &:checked ~ ${StyledRadio} {
+        border-width: ${props => `calc(${getRadioSize(props)} / 3) `};
+    }
+
+    &:checked:disabled ~ ${StyledRadio} {
+        border-color: ${({ disabledBorderColor }) => disabledBorderColor.active};
+    }
+
+    &:not(:checked):disabled ~ ${StyledRadio} {
+        border-color: ${({ disabledBorderColor }) => disabledBorderColor.default};
+    }
+
+    &:not(:disabled):checked ~ ${StyledRadio} {
+        border-color: ${({ borderColor, hasError }) => borderColor[hasError ? 'error' : 'active']};
+    }
+
+    &:not(:disabled):focus ~ ${StyledRadio} {
+        border-color: ${({ borderColor, hasError }) => borderColor[hasError ? 'error' : 'active']};
+        box-shadow: ${({ borderColor, hasError }) => `0 0 0.8rem ${borderColor[hasError ? 'error' : 'active']}`};
     }
 `;
 
-export const RadioWrapperStyled = styled('div')<RadioWrapperProps>`
-    width: ${({ theme, size }) => (size ? theme.radio.sizes[size] : theme.radio.defaultSize)};
-    height: ${({ theme, size }) => (size ? theme.radio.sizes[size] : theme.radio.defaultSize)};
+export const RadioWrapper = styled('div')<WrapperProps>`
+    margin: 0.3rem;
+    width: ${getRadioSize};
+    height: ${getRadioSize};
     position: relative;
     border-radius: 100%;
-    border: 1px solid ${({ theme }) => theme.radio.borderColor};
-    background-color: ${({ theme, disabled }) => (disabled ? theme.radio.disabledBgcolor : theme.radio.bgColor)};
-
-    &:focus-within {
-        box-shadow: 0 0 0 1pt ${({ theme }) => theme.radio.outlineColor};
-    }
-
-    ${centerAligned()}
 `;
-RadioWrapperStyled.defaultProps = {
-    theme: defaultTheme
-};
+
+export const RadioWithLabelWrapper = styled('label').attrs(({ theme }) => ({ ...theme.radio }))<WrapperProps>`
+    display: ${({ fullWidth }) => (fullWidth ? 'flex' : 'inline-flex')};
+    cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
+
+    ${getSelectorLabelPositionStyle}
+
+    &:hover {
+        && ${StyledRadio} {
+            border-color: ${({ disabled, hasError, hoveredBorderColor }) => !disabled && hoveredBorderColor[hasError ? 'error' : 'active']};
+        }
+    }
+`;