+++ packages/core/src/components/RadioGroup/RadioGroup.tsx
@@ -1,61 +1,117 @@
-import { WithStyle } from '@medly-components/utils';
-import React, { FC, useCallback } from 'react';
-import FieldWithLabel from '../FieldWithLabel';
+import { useCombinedRefs, WithStyle } from '@medly-components/utils';
+import React, { FC, useCallback, useMemo, useState } from 'react';
 import Radio from '../Radio';
+import { SelectorGroup } from '../Selectors';
 import { Props } from './types';
 
 export const RadioGroup: FC<Props> & WithStyle = React.memo(
     React.forwardRef((props, ref) => {
         const {
+            id,
             size,
+            value,
             name,
             label,
+            disabled,
             required,
             options,
-            labelPosition,
-            onChange,
-            value,
-            disabled,
             labelVariant,
-            fullWidth,
             labelWeight,
-            labelColor
+            errorText,
+            helperText,
+            columns,
+            validator,
+            onChange,
+            onBlur,
+            onInvalid,
+            ...wrapperProps
         } = props;
 
-        const handleOnChange = useCallback(
-            (event: React.FormEvent<HTMLInputElement>) => {
-                const target = event.target as HTMLInputElement;
-                onChange && onChange(target.value);
-            },
-            [onChange]
+        const [builtInErrorMessage, setErrorMessage] = useState(''),
+            radioGroupId = useMemo(() => id || name, [id, name]),
+            radioGroupRef = useCombinedRefs<HTMLDivElement>(ref, React.useRef(null)),
+            hasError = useMemo(() => !!errorText || !!builtInErrorMessage, [errorText, builtInErrorMessage]),
+            hasHelperOrErrorText = useMemo(() => !!(errorText || helperText), [errorText, helperText]);
+
+        const validate = useCallback(
+            (selectedValue: string, validationMessage?: string) =>
+                setErrorMessage((validator && validator(selectedValue)) || validationMessage),
+            [validator]
         );
+
+        const handleOnInvalid = useCallback(
+                (event: React.FormEvent<HTMLInputElement>) => {
+                    validate(value, (event.target as HTMLInputElement).validationMessage);
+                    onInvalid && onInvalid(event);
+                },
+                [validate, onInvalid, value]
+            ),
+            handleOnBlur = useCallback(
+                (event: React.FocusEvent<HTMLDivElement>) => {
+                    const currentTarget = event.currentTarget,
+                        target = event.target as HTMLInputElement;
+                    setTimeout(() => !currentTarget.contains(document.activeElement) && validate(value, target.validationMessage), 0);
+                    onBlur && onBlur(event);
+                },
+                [validate, onBlur, value]
+            ),
+            handleOnChange = useCallback(
+                (event: React.ChangeEvent<HTMLInputElement>) => {
+                    const value = event.target.value;
+                    validate(value);
+                    onChange && onChange(value);
+                },
+                [validate]
+            );
+
         return (
-            <FieldWithLabel id={`${label}-radioGroup`} ref={ref} fullWidth {...{ fullWidth, labelPosition }}>
+            <SelectorGroup.Wrapper
+                id={`${radioGroupId}-wrapper`}
+                ref={radioGroupRef}
+                type="radio"
+                aria-describedby={`${radioGroupId}-helper-text`}
+                onBlur={handleOnBlur}
+                onInvalid={handleOnInvalid}
+                {...{ ...wrapperProps, hasHelperOrErrorText }}
+            >
                 {label && (
-                    <FieldWithLabel.Label {...{ required, labelPosition, labelVariant, labelWeight, labelColor }}>
+                    <SelectorGroup.Label
+                        type="radio"
+                        hasError={hasError}
+                        id={`${radioGroupId}-label`}
+                        textVariant={labelVariant}
+                        textWeight={labelWeight}
+                        {...{ required, disabled }}
+                    >
                         {label}
-                    </FieldWithLabel.Label>
+                    </SelectorGroup.Label>
                 )}
-                <FieldWithLabel.Field onChange={handleOnChange}>
+                {(hasError || helperText) && (
+                    <SelectorGroup.HelperText id={`${radioGroupId}-helper-text`} type="radio" {...{ disabled, hasError }}>
+                        {errorText || builtInErrorMessage || helperText}
+                    </SelectorGroup.HelperText>
+                )}
+                <SelectorGroup.Options id={`${radioGroupId}-options`} columns={columns}>
                     {options.map(option => (
                         <Radio
+                            id={`${option.label}-${radioGroupId}`}
                             key={option.value}
-                            {...{ ...option, size, name, required }}
-                            disabled={disabled || option.disabled}
-                            defaultChecked={option.value === value}
+                            onChange={handleOnChange}
+                            checked={option.value === value}
+                            {...{ ...option, disabled: disabled || option.disabled, name, size, required, hasError }}
                         />
                     ))}
-                </FieldWithLabel.Field>
-            </FieldWithLabel>
+                </SelectorGroup.Options>
+            </SelectorGroup.Wrapper>
         );
     })
 );
 
 RadioGroup.displayName = 'RadioGroup';
-RadioGroup.Style = FieldWithLabel.Style;
+RadioGroup.Style = SelectorGroup.Wrapper;
 RadioGroup.defaultProps = {
-    label: '',
-    name: '',
+    columns: 1,
     fullWidth: true,
-    labelPosition: 'left'
+    labelWeight: 'Medium',
+    labelVariant: 'body1'
 };