+++ packages/core/src/components/Tabs/TabPanel/TabPanel.styled.tsx
@@ -6,4 +6,5 @@ export const TabPanel = styled('div')`
     }
     padding: 10px;
     height: 100%;
+    width: 100%;
 `;