+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://github.com/medly/medly-components/compare/@medly-components/icons@1.0.5...@medly-components/icons@1.0.6) (2020-06-29)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
+## [1.0.5](https://github.com/medly/medly-components/compare/@medly-components/icons@1.0.4...@medly-components/icons@1.0.5) (2020-06-24)
+
+**Note:** Version bump only for package @medly-components/icons
+
+
+
+
+
 ## [1.0.4](https://github.com/medly/medly-components/compare/@medly-components/icons@1.0.3...@medly-components/icons@1.0.4) (2020-06-19)
 
 **Note:** Version bump only for package @medly-components/icons